Continued work on inventory
* added better add logic, as they now try to add to all slots / fill the inventory and return the added amount
* added checks for max weight and volume including calculation and updating on insert
* changed amount of inventoryable to int, as we never used it as a decimal anyway and working with integers is easier
* made remove to remove from multiple slots and resprect fixed slot size inventories

+++ OctoAwesome/OctoAwesome/IInventorySlot.cs
@@ -14,7 +14,7 @@ namespace OctoAwesome
         /// <summary>
         /// The volume amount of <see cref="Item"/> in this slot[dm³].
         /// </summary>
-        decimal Amount { get; }
+        int Amount { get; }
         /// <summary>
         /// Gets the definition for the <see cref="Item"/>.
         /// </summary>
